[Design]#40 - 리스트 ui 수정
